refactor(client): 디렉토리 명 변경
