+++ client/internal/build.gradle
@@ -2,6 +2,6 @@ subprojects {
     apply plugin: 'java-library'
     dependencies {
         api project(':client:mathrank-client-config')
-        api project(':client:mathrank-client-result')
+        api project(':client:mathrank-client-response')
     }
 }
