+++ client/internal/mathrank-course-client/src/main/java/kr/co/mathrank/client/internal/course/CourseClient.java
@@ -11,7 +11,7 @@ import org.springframework.web.client.RestClient;
 import jakarta.validation.constraints.NotNull;
 import kr.co.mathrank.client.config.RestClientResponseDecorator;
 import kr.co.mathrank.client.config.TimeoutConfiguredClient;
-import kr.co.mathrank.client.result.ClientResponse;
+import kr.co.mathrank.client.response.ClientResponse;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
