+++ client/mathrank-client-response/src/main/java/kr/co/mathrank/client/response/ClientException.java
@@ -1,4 +1,4 @@
-package kr.co.mathrank.client.result;
+package kr.co.mathrank.client.response;
 
 public class ClientException extends RuntimeException {
 	public ClientException(String message) {
