+++ client/mathrank-client-response/src/main/java/kr/co/mathrank/client/response/ClientTimeoutException.java
@@ -1,4 +1,4 @@
-package kr.co.mathrank.client.result;
+package kr.co.mathrank.client.response;
 
 public class ClientTimeoutException extends ClientException {
 	public ClientTimeoutException(String message) {
